backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2

# OCR
pytesseract==0.3.10
easyocr==1.7.1

# Metadata & File Analysis
exifread==3.0.0
python-magic==0.4.27

# Machine Learning (Optional)
scikit-learn==1.3.2

# Utilities
aiofiles==23.2.1

